refactor(compiler): move `start` to its own file
